pyproject.toml
[project]
name = "cv-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.11",
    "openai>=2.16.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.50.0",
]
